refactor server.go
